protocol: add protocol flags and events with simple encoder.
This can either compose to a buffer or use iovec directly with writev().
In the most simple case, writev() is just barely faster than buffering.
In the pathological case, where much buffer resizing will occur, it is
a lot faster and easier on the memory allocator.

This should allow for building both a sync and async client eventually.