cluster: make sure we use db.$cmd for getlasterror.
This seemed to have been introduced when changing to the sendv() design
for pipelining write concerns. The DB_AND_CMD_FROM_COLLECTION macro will
place into a stack allocated structure the version of the namespace for
the command.

However, this will get overwritten if we have multiple RPCs comming in.
We will fix this in a followup commit that will remove sendv() support
since we do not actually use it (nor need it).