client-pool: add mongoc_client_pool_try_pop().
This can be used to request a mongoc_client_t without waiting for a new
client to be given back to the pool. If no more connections are available
then NULL is returned.

This may be advantageous in a scenario where you want to queue an incoming
HTTP request until a connection is available to service the request.