database: add mongoc_database_has_collection().
This function can be used to check to see if the MongoDB server knows
about a particular collection name.

Be careful using this, however, as it could be racey. Another client
could race to insert the first document into a collection, etc.

This is performed by sending a query to "system.namespaces".

If FALSE is returned, you should also check to see if error.domain is set
to disambiguate between a failure and no-existance.