bulk insert and IOVEC_ARRAY_FIELD rpc type
added in mongoc_collection_insert_bulk (that takes an array of bson_t
pointers and inserts them as part of one bulk insert.

to support this, added in a mongoc_collection_insert_bulk_raw, which
adds via a struct iovec *

to support that, added in a IOVEC_ARRAY_FIELD rpc type, which allows for
arbitrary trailing binary data

Also, added in actually reading the gle reply for inserts to catch
things like unique key violations.