cluster: add _mongoc_cluster_command_early().
This is meant for sending a command before the cluster has been
initialized. It forces the cluster to start-up and executes the
command.

The error value can be used to notify the call chain of connection
failure.

This is primarily meant for mongoc_client_warm_up() and may not be
useful in other scenarios where we have already connected.