client: warm up client to determine wire-version to use.
This will help with situations where we need to do an insert but do
not yet know what our negotiated wire version is.

Functions needing to know the wire-version should first call the
private _mongoc_client_warm_up() function.

It will either result in a failure, (and error set), or TRUE indicating
success and client->cluster.wire_version is valid.