collection: add mongoc_collection_find_and_modify() helper.
This should simplify the use case for people needing to do find_and_modify.
My hope is that we can maintain this ABI, as it is a lot simpler than
building complex bson documents with all the options.

In the future we may introduce a more "builder" style concept for executing
find and updates, but this should work for now.

Holler if you have any complaints.