collection: deprecate mongoc_collection_ensure_index()
The ensure_index() indicated, like other legacy drivers, that we were
caching the results of the index creation. We do not, and will not be
doing this in the future. (There are lots of good reasons why calling
ensure index all over the place is a bad idea).

Longer term, we may add a create_indexes() method so you can create more
than one at a time, but for now this will do.