stream-tls: mind the timeout units.
During the transition to absolute timeouts to handle clock drift the units
could have been 1000x the desired timeout value. stream functions take
a relative timeout where as we track our absolute timeout using the
monotonic clock, which is in microseconds.

This changeset adjusts the name to make it clear which units we are using
for the timeout to BIO methods.

Fixes #43