bulk: Handle upsert on MongoDB < 2.6 with non-objectid _id fields.
Server versions before 2.6 only return the 'upserted' field in an update
result when the upserted _id is an ObjectId. If the user provides any
other value for _id in the query spec, or update document (for a replace
operation) the upserted field is not returned. This causes invalid counts
for nUpserted in the Bulk API result set.

This fixes CDRIVER-372.