client: translate slaveok into secondary preferred.
This will allow clients to dispatch to secondaries for reads if
slaveok=true in the MongoDB connection URI as the default read
preference.

We might want to determine if SECONDARY_PREFERRED is the ideal, or
if we want NEAREST. Since NEAREST is a bit confusing in how it actually
works, SECONDARY_PREFERRED is probably the way to go.