Conditionally compile SSL and SASL
This allows building *.[ch] even when libssl and/or libsasl are not available.

This helps get the library to build for iOS, and similarly would be helpful for building the library under any third-party build system.

Sort of an aside: on Mac OS libssl is deprecated. iOS has never shipped with it. On Mac OS the headers are in /usr/include but aren't available when building an iOS project.

Ideally, someday, libmongoc would support both libssl and Common Crypto, Apple's implementation.