CDRIVER-570: Executing bulk without write concern should inherit from client
When using mongoc_bulk_operation_new() to construct a bulk operation (not through a collection), one would expect the client's write concern to be used at execution time if an explicit write concern is not set on the bulk operation. That functionality is already handled by _mongoc_write_command_execute(), but it requires we not initialize a new write concern on the bulk operation.