CDRIVER-642 detect auth error from usersInfo
Starting with MongoDB 2.7.1 the scope of the localhost exception for
authentication was dramatically narrowed. It is no longer possible to call the
userInfo command unauthenticated, even when the localhost exception is in
effect. A workaround for this issue is to catch the exception and call
createUser if the error code is 13 (Unauthorized). Any exception from the
createUser call should propagate to the user application.