CDRIVER-631 reduce max retries from 3 to 1
It can take a surprising amount of time before we report primary not available
when connecting to a replica set with two downed members. With
socketTimeoutMS = connectionTimeoutMS = 1 second, it can take up to 16 seconds
to return an error from an operation.

Reducing retries to 1 speeds this scenario to bail after 8 seconds, not 16.