generate mock_server2 functions with python, not macros
I'd hacked up macros to generate functions for getting and setting
values of various types for the future_t implementation, and for
generating functions like future_cursor_next.

The macro magic / hackery was very hard to understand, but it did (just
barely) work in GCC and Clang. When it broke in MSVC, however, I saw I
was in over my head. Generating the functions with Python is clear and
maintainable.