CDRIVER-678: Only the first member of seedlist is used
Previously we would only add the node (topology description) to the
topology scanner when extracting it from a seedlist.
This resulted in no new nodes were ever added/connected to -- we would
only ever connect to the nodes in the seedlist -- no actual topology
discovery was being done.

Now we make sure we add the node to the scanner when we get a new topology
description, and start monitoring it.
When we have new ismaster results, we add new nodes, and remove old
nodes to/from the scanner.