CDRIVER-742 disable mock server tests on Solaris
get_port() in mock_server.c often fails on Solaris (tested on Solaris 11
specifically in Evergreen), logging "Failed to get listening port number:
Socket operation on non-socket". This leads to test failures so I've disabled
the mock server tests on Solaris.

It would be nice to determine what's wrong with the mock server test framework
on Solaris and reënable the tests.