CDRIVER-770: Return how much we did write since its OK to return short writes
When we can't finish writing our entire buffered iovec, we must still
return how much we have successfully written so far, in case any of
the previous buffers finished.
Then we can regroup and recalculate from where we left off and retry in
case of an [WSA]EWOULDBLOCK or friends.