CDRIVER-846 refactor insert commands
_mongoc_write_command_init_insert and _mongoc_write_command_insert_append
differ unnecessarily from the equivalent helper functions for update and
remove: the insert functions take an array of documents instead of a single
document.

(The inconsistency dates from the implementation of
mongoc_collection_insert_bulk, now deprecated. In any case,
mongoc_collection_insert_bulk can be responsible for handling an array of
documents instead of pushing that chore down the call stack.)