CDRIVER-830 do blocking reconnect if single-threaded
The new mongoc_cluster_node_reconnect initiates a non-blocking reconnect
but doesn't wait for connection to complete. If the driver is in single-
threaded mode (so the scanner and the application share a connection)
and immediately uses that server, connection hasn't completed:

cursor error: Failure during socket delivery: Socket not connected (57)

With this change, apply connectTimeoutMS to the reconnection in single-
threaded mode.