CDRIVER-906 move read pref & error logic from mongoc-cursor.c
Preparing to separate command path from query path. No new code, just
moving from mongoc-cursor.c the code that will be shared between
commands and queries: the logic to apply a read preference to an
OP_QUERY, and the logic to turn an OP_REPLY into a bson_error_t.