CDRIVER-935 pooled clients require SSL if opts set
In 1.1.x, all clients (pooled and single-threaded) required an SSL
connection to the server only if "ssl=true" was in the URI. In 1.2.0
this changed unintentionally (but beneficially) for single-threaded
clients: setting a single-threaded client's SSL options put the client
in SSL mode, even if "ssl=true" is omitted from the URI.

Pooled clients, however, stopped working entirely if the options are
set but "ssl=true" is omitted. This patch fixes pooled clients and
tests the new behavior for both.