CDRIVER-982 set OP_QUERY nToReturn from limit
The "limit" argument to mongoc_collection_find should be used for two purposes:

1. Mark a mongoc_cursor_t as "done" once it has fetched "limit" documents

2. Send limit to the server as the nToReturn field in the OP_QUERY and
OP_GETMORE message headers

The driver respects #1, but not #2. Instead, it always sends nToReturn of 0 to
the server, meaning, "accept server default" batch size of 101 documents at
first, then 4MB of documents after. This bug incurs a performance penalty: the
final batch is always full-sized (about 4MB) even if the cursor only needs part
of it. If mongoc_collection_find's batch_size is non-0 the logic is correct;
it's when batch_size is 0 (the default) that limit is incorrectly ignored when
formatting OP_QUERY and OP_GETMORE.