CDRIVER-1117 delete mongoc_cluster_run_command_rpc
I introduced the function for CDRIVER-838 to implement "find"/"getMore"
commands -- the cursor was built to send and receive RPCs, so it was
easier to stick a find or getMore command into an RPC and send it,
instead of using the standard command path from the cursor code.

This causes some duplication, however, and it makes it very hard to
consistently record command events originating from cursors and those
that don't (CDRIVER-769).

Remove mongoc_cluster_run_command_rpc and use the standard command path.
When a cursor runs a command, it no longer stores the reply in its rpc
struct; instead it stores the reply in a bson_t on its iface_data.