CDRIVER-1157: Verify certificates during handshake
This simplifies the OpenSSL code a lot, while this was already
the case in Secure Transport

Note that the previous functions have been deprecated and
always return false now.
We cannot remove them due to ABI, but there is no chance
a user of the driver would have been calling them intentionally
since he would have to get a hold of the stream between
us creating it, and before we even run ismaster on it.