CDRIVER-1179 don't disable flags g++ rejects
In ./autogen.sh --enable-maintainer-flags, for some reason we check if
both gcc and g++ accept each flag, and disable the flag if not. Thus we
disabled -Wdeclaration-after-statement, which is a useful flag to have
turned on when compiling C.

We don't use g++, so don't check flags for g++ compatibility.