CDRIVER-1407 fix socket test on FreeBSD
Fixes a timeout in /Socket/sendv test.

The test had wrongly assumed that a sendv timeout set socket->errno_,
but this errno was actually set from the initial call to accept or
connect. These functions do not set socket->errno_ during this test of
localhost connections on FreeBSD, the way they do on Mac and Linux.