CDRIVER-2083 use target_compile_definitions
add_definitions() has global effect, so isn't good to include
in a .cmake file.  This commit instead publishes the definitions in a
variable and lets consumers set it on a target the same way they set
library/include paths.

mongocxx (and maybe the PHP driver) will need a corresponding update.