CDRIVER-2172 check idle socket with "ping"
Server Selection Spec update for single-threaded drivers.

If a server is selected that has an existing connection that has been idle for
socketCheckIntervalMS, the driver MUST check the connection with the "ping"
command. If the ping succeeds, use the selected connection. If not, set the
server's type to Unknown and update the Topology Description according to the
Server Discovery and Monitoring Spec, and attempt once more to select a server.