CDRIVER-573 "not master" err invalidates server
Server Discovery and Monitoring Spec: a "not master" or "node is
recovering" error from any command marks the server Unknown.

This has been a missing feature since we implemented SDAM in libmongoc
1.2.0. It's now urgent to fix because Retryable Writes depends on this
feature. Otherwise during a stepdown, the driver might try a write
against a former primary, get a "not master" error, and try again with
the same server. Marking it Unknown ensures we re-scan the topology.