CDRIVER-2293 add snappy/zlib flags to config files
When we bundle zlib and/or snappy, currently the Autotools build system
actually builds and installs the libraries alongside libmongoc. To link
statically to libmongoc in this configuration, you need to include the
full path to zlib and/or snappy in the link line.

This patch adds the full path to zlib/snappy in the CMake config file.

This is a temporary workaround until we fully vendor zlib/snappy in
CDRIVER-2363.