CDRIVER-2228 don't cache txnNumber iter
We had attempted to optimize by keeping a bson_iter_t pointed to the
location of the txnNumber field within bulk write commands, but there's
some bug with this: intermittently we send a command to the server with
txnNumber reset to 0. Let's not do this optimization now, and revisit
later.