CDRIVER-2214 prohibit read/writeConcern w/ old server
Commands that write now conform to the drivers spec: it is an error to
supply an explicit write concern if MongoDB is too old to support it.

Commands that read now conform to the spec as well: it is an error to
supply an explicit read concern *or* to inherit a read concern from the
client, database, or collection, if MongoDB is too old to support read
concern.