CDRIVER-2578 changestream getmore and agg lsid
collection_watch runs an aggregate command with
collection_read_command_with_opts and then creates a cursor with
cursor_new_from_command_reply. To ensure they use the same implicit
session lsid, the change_stream_t must be responsible for creating
the implicit session and pass this session to both the command and
cursor as if it were an explicit session.