travis: move osx to its own matrix include
Rather than putting the Linux only or exceptions everywhere else.

Add linux-ppc64le as a test. Because the travis ppc64le test
infrastructure is based on xenial the distro mongodb is 2.x which isn't
suitable for the test. Pulling in 3.x from ubuntu bionic resulted in
conflicts so we use the enterprise mongo for ppc64le xenial to have
a workable test server for the client C implementation.