CDRIVER-3340 permit NULL as binary value to append
Copying a binary value with 0 length would result in a
bson_value_t with a NULL value for the binary data, causing
a crash if that copied value was appended. This was due to
bson_malloc returning NULL if allocating 0 bytes. So permit
NULL as a binary value to append.