CDRIVER-3678 fix /Topology/request_scan_on_error
CDRIVER-3615 changed an exotic edge case for parsing errors.
An error with a NotMaster error code, but a "node is recovering" message
is considered a "node is recovering" error for multi-threaded error
handling. So requested scan is expected.

A "node is recovering" error should request scan for multi-threaded, not
single-threaded scanning. The test was making a wrong assertion, but the
check of the assertion was very racy.

This fixes the test and increases the wait to make the test less racy.