Set Certain Warnings as Errors Unconditionally (#890)
Some constructs that are allowed in C are well-established to be virtually always erroneous, e.g. calling a function that hasn't been declared, implicitly binding a pointer to a different type, and implicitly converting an integral to a pointer. Implicitly converting a pointer-to-const to a pointer-to-mutable is not always an error, but is problematic enough that an explicit cast should be required.

Many of these errors will be caught at a later time, but only due to test failures, linker failures, or on a build on CI that enables some warnings-as-errors (currently never on Windows).

This change instead makes such constructs always-errors, thus adding a barrier to any changes that produce such errors that might otherwise have been missed. The selected warnings are well-established and detected by all modern compilers, and are very unlikely to appear anew do to compiler upgrades.

Additional warnings that are not so definitely-wrong (etc. sign conversion/compare, unused variables) are not addressed by these changes.

* MSVC fixes

* More warnings, and fixes for C-only warnings

* Only enable warnings in certain subdirectories

* Include mongoc-warnings in dist

* Disable some warnings on GCC < 5.0

* Use an OBJECT library for bundled zlib

* Disable compiler warnings for zlib

* Document MaintainerFlags versus MongoC-Warnings