fix build of bundled zlib
Ensure that the target for the bundled zlib is only built when the
bundled target is needed.  Also ensure that the build always correctly
finds the necessary headers.  For instance, even if there is a zlib
installed on the system and the build specifies the bundled zlib, ensure
that bundled headers are located before the system headers.