CDRIVER-3973: Support the ignoreExtraEvents property (#1012)
This doesn't fully complete the tasks of CDRIVER-3973, but finishes a
single isolated subtask thereof: Supprt for the `ignoreExtraEvents`
property on `expectedEventsForClient`, which is a prereq for enabling
the new changestreams tests.

In this case, if `ignoreExtraEvents` is specified as `true`, we won't
generate an error if the "actual events" contains more events than
"expected events." Each expected event is still searched for within
the "actual events" list, and anything additional is ignored.