Faster and Simpler Warning Flag Handling (#1318)
* Move maintainer flags into CMakeLists directly

This puts the optional compile warnings in CMake code directly and
applies them using generator expressions. This also speeds up subsequent
configure times substantially as it doesn't run repeated try_compiles

* Fix: alignof() detection warning when compiling as C++

We check for __STDC_VERSION__ which may be undefined for C++ and on
older C compilers. Guard against C++ and check that it is defined before
considering it. (Fixes -Wundef warnings)

* Mark SSL header-dirs as SYSTEM

The OpenSSL/LibreSSL headers tricker redundant-decl warnings, so we
want to mark those as "system", which will disable warnings for the
content of those headers.

* Fix Wformat error with type mismatch.

Formatting with PRIu8 expects a uint8_t, but we pass plain
(unsigned) int. This only appears to warn on some compilers? Fix this
with an explicit cast.