Fix testing instructions
Previously, following the testing instructions in CONTRIBUTING.md
results in the following error:
```
"error":"NonExistentPath: Data directory /data/db not found. Create the missing directory or specify another path using (1) the --dbpath command line option, or (2) by adding the 'storage.dbPath' option in the configuration file."
```
This adds the dbpath to the command line so that way one can copy and
paste the instructions in order to run tests.

Also, prefer to use the newer mongosh over mongo.