Add an in-source example/test that uses vcpkg
This is a very minimal example and test case that installs the
release (not local!) version of libbson and libmongoc, and then tests
that the application builds and runs correctly. There is no CI task
for this yet, as it is unstable and does not make use of the local
source tree. Future updates may make a more thorough test case that
overrides the source for the vcpkg mongo-c-driver port to use the local
source tree, thus giving a route for automated testing of changes
against the third-party portfile.