[CDRIVER-4651] Documentation on vcpkg installation + more (#1424)
* Docs additions and updates:

- Enable the documentation index. Begin using index entries.
- Add a reference page for Linux, Conand, and vcpkg packages and how
  they can be found/used.
- Add a tutorial section for installing via vcpkg.

* Add an in-source example/test that uses vcpkg

This is a very minimal example and test case that installs the
release (not local!) version of libbson and libmongoc, and then tests
that the application builds and runs correctly. There is no CI task
for this yet, as it is unstable and does not make use of the local
source tree. Future updates may make a more thorough test case that
overrides the source for the vcpkg mongo-c-driver port to use the local
source tree, thus giving a route for automated testing of changes
against the third-party portfile.

* Add links to external port+recipes
* Tweak inclusion of the genindex page.

Including the genindex in a toctree seems to create a conflict with
the full_index page which tries to include the root `index.rst` via
another toctree? This is odd, but easy enough to work around.