CDRIVER-4743 Define `BSON_STATIC` for `bson_static` build (#1453)
* remove `bson_obj`

Separate sources built by `bson_shared` and `bson_static`. This is intended to build `bson_static` sources with no `dllexport` and `bson_shared` with `dllexport`.

* define `BSON_STATIC` when building

To suppress `dllexport`. When `dllexport` is defined, it appears to export the symbols in a consumer.

* revise comments

BSON_STATIC and MONGOC_STATIC are defined for static builds.
Exporting functions in static build has an effect.