CDRIVER-4788 do not require sphinx-design for man pages (#1483)
* do not require sphinx-design

sphinx-design is not used in the man pages. Do not require sphinx-design to enable building man pages when sphinx-design is not available.

* error if building HTML docs without sphinx-design

* remove `python3-sphinx` from mock install

`python3-sphinx` is listed in BuildRequires of spec file and does not need to be installed separately.

* update spec.patch to remove `python3-sphinx-design`

Intended to test the RPM build without the `python3-sphinx-design` dependency installed.

* commit staged files

This is intended to test files changed in a patch build.
Patch builds apply changes to index.