CDRIVER-4767 support Sphinx 1.7.6 (#1492)
* import older `DirectoryHTMLBuilder` if needed

The import path of `DirectoryHTMLBuilder` was changed from `sphinx.builders.html` to `sphinx.builders.dirhtml` in Sphinx v2.0.0

* import `Directive`, not `SphinxDirective`

`SphinxDirective` was added in Sphinx v2.0.0

* do not import Project

`Project` was added in Sphinx v2.0.0

* update `needs_sphinx` to 1.7

* error if `add_css_file` is unavailable for HTML builder.

But do not error otherwise. This is intended to help build man pages without HTML requirements.

`Sphinx.add_css_file` was added in Sphinx v3.5.0

* use `builder-inited`

`config-inited` was added in Sphinx v1.8.0

* do not use `external` role

`external` role was added in intersphinx extension in v4.4

* do not use `noindexentry`

`noindexentry` option was added to C domain in Sphinx v3.2

* remove `name` from `index` directive

`name` option was added in Sphinx v3.0

* populate `man_pages` in `setup` instead of event handler

The `config-inited` event was added in Sphinx v1.8.0
The `ManualPageBuilder` constructor raises an exception for an empty `man_pages` value.

* remove parallel build of man pages

To avoid an error observed on Sphinx 1.7.6:
```
AttributeError: Can't pickle local object 'BuildEnvironment._read_parallel.<locals>.merge'
```

* remove unhelpful warning

An error is raised at runtime with a message suggesting to install `sphinx-design`. The warning was a source of confusion in CDRIVER-4767

* update NEWS for 1.25.3

* add upcoming NEWS for 1.25.4