CDRIVER-4663 Eagerly create minPoolSize connections
This adds a new public function to the libmongoc API.

mongoc_client_pool_warmup "warms up" aka eagerly connects the specified
number of clients in a client pool to the MongoDB servers in the pool's
topology.

This function is thread safe and blocks until the pool contains
'num_to_warmup' clients.