unified tests: serverDescriptionChangedEvent, and refactor event serialization
The command-logging-and-monitoring tests use the serverDescriptionChanged event. In implementing a new event type, it made sense to refactor the implementation for event storage and serialization. The serialization from observeEvents and storeEventsAsEntities unnecessarily diverged into a late-serialization and an early-serialization style. This decides on early serialization, opting to build a canonical bson representation as soon as possible and extract fields from it later as necessary for comparisons. This has fewer opportunities for optimization in some cases, but here the extra simplicity is a good match for the test runner environment.