Let's start a trend of making bson serialization a shared internal API
Serialization will be needed for other data types too, and this seems better than having each consumer do their own bson serialization. This design appends fields to an existing bson document.

Chose the name to be distinct from 'serialize' (which is vague, and often used in a concurrency context in libmongoc) and 'as_bson' which is not clear about whether the document is being allocated or not.