example-structured-log: motivate the thread safety with a global counter
The printf() is meant to stand in for a stream of some sort, but on its own doesn't motivate thread safety because stdio has its own locking.

We could use a non-threadsafe file or queue, but Kevin suggested a counter and I like how simple this is. The code now includes a counter, and the comments mention that it could represent something mroe complex like a file or queue.